services/data_collector/requirements.txt
# services/data_collector/requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
python-dotenv
yfinance
numpy
pandas
redis
aiohttp
boto3
elasticsearch
watchtower
psutil